+++ src/components/controlled/transactions-table/components/transaction-hash/transaction-hash.tsx
@@ -4,11 +4,13 @@ import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 import type { TransactionRowType } from '../../transactions-table.type';
 
 const transactionHashClasses: Record<string, string> = {
-  explorerLink: 'mvx:text-blue-link!',
+  explorerLink: 'mvx:text-primary!',
+  transactionHash: 'mvx:flex mvx:items-center mvx:justify-center',
 };
 
 @Component({
   tag: 'mvx-transaction-hash',
+  styleUrl: 'transaction-hash.scss',
 })
 export class TransactionHash {
   @Prop() class?: string;
@@ -20,15 +22,20 @@ export class TransactionHash {
     }
 
     return (
-      <div class={{ 'transaction-hash': true, [this.class]: Boolean(this.class) }}>
-        <mvx-transaction-icon iconInfo={this.transaction.iconInfo} />
+      <div
+        class={{
+          'transaction-hash': true,
+          [this.class]: Boolean(this.class),
+        }}
+      >
+        <mvx-transaction-icon iconInfo={this.transaction.iconInfo} class={transactionHashClasses.transactionHash} />
 
         <mvx-explorer-link
           dataTestId={DataTestIdsEnum.transactionLink}
           link={this.transaction.link}
           class={transactionHashClasses.explorerLink}
         >
-          <span>{this.transaction.txHash}</span>
+          <mvx-trim text={this.transaction.txHash} />
         </mvx-explorer-link>
       </div>
     );