+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Updated styles on transaction table](https://github.com/multiversx/mx-sdk-dapp-ui/pull/181)
+
 ## [[0.0.22](https://github.com/multiversx/mx-sdk-dapp-ui/pull/179)] - 2025-08-01
 
 - [Fixed side panel overflow and testids](https://github.com/multiversx/mx-sdk-dapp-ui/pull/178)